fix image drop position
