+++ ui/editor/plugins/upload-images.tsx
@@ -1,6 +1,7 @@
-import { handleImageUpload } from "@/lib/editor";
-import { Plugin, PluginKey } from "@tiptap/pm/state";
-import { Decoration, DecorationSet } from "@tiptap/pm/view";
+import { BlobResult } from "@vercel/blob";
+import { toast } from "sonner";
+import { EditorState, Plugin, PluginKey } from "@tiptap/pm/state";
+import { Decoration, DecorationSet, EditorView } from "@tiptap/pm/view";
 
 const uploadKey = new PluginKey("upload-image");
 
@@ -43,13 +44,15 @@ const UploadImagesPlugin = () =>
     },
   });
 
-function findPlaceholder(state, id) {
+export default UploadImagesPlugin;
+
+function findPlaceholder(state: EditorState, id: {}) {
   const decos = uploadKey.getState(state);
   const found = decos.find(null, null, (spec) => spec.id == id);
   return found.length ? found[0].from : null;
 }
 
-export function startImageUpload(file, view) {
+export function startImageUpload(file: File, view: EditorView, pos: number) {
   // A fresh object to act as the ID for this upload
   const id = {};
 
@@ -63,7 +66,7 @@ export function startImageUpload(file, view) {
     tr.setMeta(uploadKey, {
       add: {
         id,
-        pos: tr.selection.from,
+        pos,
         src: reader.result,
       },
     });
@@ -93,4 +96,53 @@ export function startImageUpload(file, view) {
   });
 }
 
-export default UploadImagesPlugin;
+export const handleImageUpload = (file: File) => {
+  // check if the file is an image
+  if (!file.type.includes("image/")) {
+    toast.error("File type not supported.");
+
+    // check if the file size is less than 50MB
+  } else if (file.size / 1024 / 1024 > 50) {
+    toast.error("File size too big (max 50MB).");
+  } else {
+    // upload to Vercel Blob
+    return new Promise((resolve) => {
+      toast.promise(
+        fetch("/api/upload", {
+          method: "POST",
+          headers: {
+            "content-type": file?.type || "application/octet-stream",
+            "x-vercel-filename": file?.name || "image.png",
+          },
+          body: file,
+        }).then(async (res) => {
+          // Successfully uploaded image
+          if (res.status === 200) {
+            const { url } = (await res.json()) as BlobResult;
+            // preload the image
+            let image = new Image();
+            image.src = url;
+            image.onload = () => {
+              resolve(url);
+            };
+            // No blob store configured
+          } else if (res.status === 401) {
+            resolve(file);
+
+            throw new Error(
+              "`BLOB_READ_WRITE_TOKEN` environment variable not found, reading image locally instead.",
+            );
+            // Unknown error
+          } else {
+            throw new Error(`Error uploading image. Please try again.`);
+          }
+        }),
+        {
+          loading: "Uploading image...",
+          success: "Image uploaded successfully.",
+          error: (e) => e.message,
+        },
+      );
+    });
+  }
+};
